test split files
